requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.3.0
asgi-csrf==0.9
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.1.0
bitarray==2.8.1
blinker==1.7.0
boltons==21.0.0
bracex==2.4
cattrs==23.2.3
cbor2==5.4.6
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-default-group==1.2.4
click-option-group==0.5.6
colorama==0.4.6
crytic-compile @ git+https://github.com/crytic/crytic-compile.git@b5c538aaa66be44b7a68d9723881a7eba2c20898
cytoolz==0.12.2
datasette==0.64.6
datasette-query-history==0.2.3
defusedxml==0.7.1
elastic-transport==8.10.0
eth-abi==4.2.1
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.1
evm-cfg-builder==0.3.1
exceptiongroup==1.2.0
face==22.0.0
Flask==3.0.0
frozenlist==1.4.0
future==0.18.3
gitdb==4.0.11
GitPython==3.1.43
glom==22.1.0
gunicorn==21.2.0
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.5
httpx==0.27.0
hupper==1.12.1
idna==3.4
itsdangerous==2.1.2
janus==1.0.0
Jinja2==3.1.2
jsonify==0.5
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
lsprotocol==2023.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mergedeep==1.3.4
multidict==6.0.4
numpy==1.26.0
packaging==23.1
pandas==2.1.1
parsimonious==0.9.0
peewee==3.17.1
Pint==0.23
pluggy==1.4.0
prettytable==3.9.0
protobuf==4.24.3
pycryptodome==3.19.0
pyevmasm==0.2.3
pygls==1.3.1
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.3.post1
pyunormalize==15.0.0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.7.0
rlp==3.0.0
rpds-py==0.10.3
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
semantic-version==2.10.0
semgrep==1.61.1
six==1.16.0
slither-analyzer==0.10.2
slither-lsp==0.0.1
slitherin==0.6.0
smmap==5.0.1
sniffio==1.3.1
solc-select==1.0.4
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.29.0
wcmatch==8.5
wcwidth==0.2.6
web3==6.9.0
websockets==11.0.3
Werkzeug==3.0.1
yarl==1.9.2
